requirements/dev.lock.txt
Jinja2==3.1.6
MarkupSafe==3.0.3
PyMuPDF==1.27.1
PyYAML==6.0.3
Pygments==2.19.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
eval_type_backport==0.3.1
fastapi==0.133.1
ffmpy==1.0.0
filelock==3.24.3
frozenlist==1.8.0
fsspec==2026.2.0
google-auth==2.48.0
google-genai==1.64.0
googleapis-common-protos==1.72.0
gradio==6.6.0
gradio_client==2.1.0
griffe==1.15.0
greenlet==3.3.2
groovy==0.1.2
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httpx-sse==0.4.3
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
invoke==2.2.1
jiter==0.13.0
jsonschema-specifications==2025.9.1
jsonschema==4.26.0
markdown-it-py==4.0.0
mcp==1.26.0
mdurl==0.1.2
mistralai==1.12.4
multidict==6.7.1
numpy==2.4.2
openai==2.29.0
openai-agents==0.12.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
packaging==26.0
pandas==3.0.1
pillow==12.1.1
playwright==1.58.0
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic-settings==2.13.1
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyee==13.0.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pymongo==4.11.2
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
ruff==0.15.2
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.3.3
starlette==0.52.1
tenacity==9.1.4
tomlkit==0.13.3
tqdm==4.67.3
typer-slim==0.24.0
typer==0.24.1
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
